reorga

+++ libft/srcs/size_len_maths/ft_negative_power.c
@@ -3,18 +3,18 @@
 /*                                                        :::      ::::::::   */
 /*   ft_negative_power.c                                :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: user42 <user42@student.42.fr>              +#+  +:+       +#+        */
+/*   By: nfauconn <nfauconn@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/23 15:42:03 by user42            #+#    #+#             */
-/*   Updated: 2022/06/23 15:42:28 by user42           ###   ########.fr       */
+/*   Updated: 2022/11/03 14:44:40 by nfauconn         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-float	ft_negative_power(int nb, int power)
+double	ft_negative_power(int nb, int power)
 {
-	float	res;
+	double	res;
 
 	power *= -1;
 	res = 1 / ft_power(nb, power);
